+++ bibigrid-core/src/main/java/de/unibi/cebitec/bibigrid/core/intents/CreateCluster.java
@@ -12,6 +12,7 @@ import java.util.Base64;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.yaml.snakeyaml.Yaml;
 
 import java.io.*;
 import java.nio.charset.StandardCharsets;
@@ -32,8 +33,8 @@ import static de.unibi.cebitec.bibigrid.core.util.VerboseOutputFilter.V;
 public abstract class CreateCluster extends Intent {
     private static final Logger LOG = LoggerFactory.getLogger(CreateCluster.class);
     public static final String PREFIX = "bibigrid-";
-    public static final String MASTER_NAME_PREFIX = PREFIX + "master";
-    public static final String SLAVE_NAME_PREFIX = PREFIX + "slave";
+    static final String MASTER_NAME_PREFIX = PREFIX + "master";
+    static final String SLAVE_NAME_PREFIX = PREFIX + "slave";
 
     protected final ProviderModule providerModule;
     protected final Client client;
@@ -282,7 +283,7 @@ public abstract class CreateCluster extends Intent {
                     sshSession.disconnect();
                 }
             } catch (IOException | JSchException e) {
-                LOG.error("SSH: {}", e);
+                LOG.error("SSH: ", e);
             }
         }
         if (configured) {
@@ -292,6 +293,16 @@ public abstract class CreateCluster extends Intent {
         }
     }
 
+    /**
+     * Uploads ansible roles to master instance.
+     *
+     * @param sshSession ssh connection to master
+     * @param hostsConfig Configuration and list of slave IPs
+     * @param commonConfig common Configuration
+     * @param slaveInstances list of slave instances
+     * @return true, if upload completed successfully
+     * @throws JSchException possible SSH connection error
+     */
     private boolean uploadAnsibleToMaster(Session sshSession, AnsibleHostsConfig hostsConfig,
                                           AnsibleConfig commonConfig, List<Instance> slaveInstances) throws JSchException {
         boolean uploadCompleted;
@@ -300,8 +311,126 @@ public abstract class CreateCluster extends Intent {
         LOG.info(V, "Connecting sftp channel...");
         channel.connect();
         try {
-            // Collect the Ansible files from resources for upload
+            // Collect Ansible files from resources for upload
             AnsibleResources resources = new AnsibleResources();
+            this.uploadResourcesFiles(resources, channel);
+
+            // Divide into master and slave roles to write in site.yml
+            Map<String, String> customMasterRoles = new LinkedHashMap<>();
+            Map<String, String> customSlaveRoles = new LinkedHashMap<>();
+
+            // Add Ansible roles
+            List<Configuration.AnsibleRoles> ansibleRoles = config.getAnsibleRoles();
+            for (Configuration.AnsibleRoles role : ansibleRoles) {
+                String roleName = getSingleFileName(role.getFile()).split(".tgz")[0].split(".tar.gz")[0];
+                Map<String, Object> roleVars = role.getVars();
+                // Set role key - value pairs
+                if (role.getVarsFile() != null) {
+                    // VarsFile readable since it is proved in Validation
+                    String vars = new String(Files.readAllBytes(Paths.get(role.getVarsFile())));
+                    Yaml yaml = new Yaml();
+                    Map<String, Object> additionalVars = yaml.load(vars);
+                    roleVars.putAll(additionalVars);
+                }
+                String roleVarsFile = "";
+                if (roleVars != null && !roleVars.isEmpty()) {
+                    roleVarsFile = roleName + "-vars.yml";
+                    commonConfig.writeAnsibleVarsFile(channel.put(channel.getHome() + "/" +
+                            AnsibleResources.CONFIG_ROOT_PATH + roleVarsFile), roleVars);
+                }
+                switch (role.getHosts()) {
+                    case "master":
+                        customMasterRoles.put(roleName, roleVarsFile);
+                        break;
+                    case "slaves":
+                        customSlaveRoles.put(roleName, roleVarsFile);
+                        break;
+                    default:
+                        customMasterRoles.put(roleName, roleVarsFile);
+                        customSlaveRoles.put(roleName, roleVarsFile);
+                }
+
+                this.uploadAnsibleRole(channel, resources, role.getFile());
+            }
+
+            // Add galaxy roles
+            List<Configuration.AnsibleGalaxyRoles> ansibleGalaxyRoles = config.getAnsibleGalaxyRoles();
+            for (Configuration.AnsibleGalaxyRoles role : ansibleGalaxyRoles) {
+                String roleName = role.getName();
+                Map<String, Object> roleVars = role.getVars();
+                // Set role key - value pairs
+                // Put vars from external vars file into Map
+                if (role.getVarsFile() != null) {
+                    String vars = new String(Files.readAllBytes(Paths.get(role.getVarsFile())));
+                    Yaml yaml = new Yaml();
+                    Map<String, Object> additionalVars = yaml.load(vars);
+                    roleVars.putAll(additionalVars);
+                }
+                String roleVarsFile = "";
+                if (roleVars != null && !roleVars.isEmpty()) {
+                    roleVarsFile = roleName + "-vars.yml";
+                    commonConfig.writeAnsibleVarsFile(channel.put(channel.getHome() + "/" +
+                            AnsibleResources.CONFIG_ROOT_PATH + roleVarsFile), roleVars);
+                }
+                // Replace ansible galaxy name with self-specified
+                role.setName(roleName);
+                switch (role.getHosts()) {
+                    case "master":
+                        customMasterRoles.put(roleName, roleVarsFile);
+                        break;
+                    case "slaves":
+                        customSlaveRoles.put(roleName, roleVarsFile);
+                        break;
+                    default:
+                        customMasterRoles.put(roleName, roleVarsFile);
+                        customSlaveRoles.put(roleName, roleVarsFile);
+                }
+            }
+
+            // Write the hosts configuration file
+            try (OutputStreamWriter writer = new OutputStreamWriter(channel.put(channel.getHome() + "/" +
+                    AnsibleResources.HOSTS_CONFIG_FILE), StandardCharsets.UTF_8)) {
+                writer.write(hostsConfig.toString());
+            }
+            // Write the commons configuration file
+            commonConfig.writeCommonFile(channel.put(channel.getHome() + "/"
+                    + AnsibleResources.COMMONS_CONFIG_FILE));
+
+            // Write custom site file
+            commonConfig.writeSiteFile(channel.put(channel.getHome() + "/"
+                            + AnsibleResources.SITE_CONFIG_FILE),
+                     customMasterRoles, customSlaveRoles);
+
+            // Write requirements file for ansible-galaxy support
+            if (!ansibleGalaxyRoles.isEmpty()) {
+                commonConfig.writeRequirementsFile(channel.put(channel.getHome() + "/"
+                        + AnsibleResources.REQUIREMENTS_CONFIG_FILE));
+            }
+
+            // Write slave instance specific configuration file
+            for (Instance slave : slaveInstances) {
+                String filename = channel.getHome() + "/" + AnsibleResources.CONFIG_ROOT_PATH + "/"
+                        + slave.getPrivateIp() + ".yml";
+                commonConfig.writeInstanceFile(slave, channel.put(filename));
+            }
+            uploadCompleted = true;
+        } catch (SftpException | IOException e) {
+            LOG.error("SFTP: ", e);
+            uploadCompleted = false;
+        } finally {
+            channel.disconnect();
+        }
+        return uploadCompleted;
+    }
+
+    /**
+     * Uploads common Ansible Resources files.
+     *
+     * @param resources ansible configuration
+     * @param channel client side of sftp server channel
+     */
+    private void uploadResourcesFiles(AnsibleResources resources, ChannelSftp channel) {
+        try {
             // First the folders need to be created
             createSftpFolders(channel, resources, resources.getFiles());
             // Each file is uploaded to it's relative path in the home folder
@@ -312,59 +441,51 @@ public abstract class CreateCluster extends Intent {
                 LOG.info(V, "SFTP: Upload file {}", fullPath);
                 channel.put(stream, fullPath);
             }
-            // @ToDo JK :: Support custom roles
-//            for (int i = 0; i < config.getMasterAnsibleRoles().size(); i++) {
-//                uploadAnsibleRole(channel, resources, config.getMasterAnsibleRoles().get(i),
-//                        commonConfig.getCustomRoleName("master", i));
-//            }
-//            for (int i = 0; i < config.getSlaveAnsibleRoles().size(); i++) {
-//                uploadAnsibleRole(channel, resources, config.getSlaveAnsibleRoles().get(i),
-//                        commonConfig.getCustomRoleName("slaves", i));
-//            }
-            // Write the hosts configuration file
-            try (OutputStreamWriter writer = new OutputStreamWriter(channel.put(channel.getHome() + "/" +
-                    AnsibleResources.HOSTS_CONFIG_FILE), StandardCharsets.UTF_8)) {
-                writer.write(hostsConfig.toString());
-            }
-            // Write the commons configuration file
-            commonConfig.writeCommonFile(channel.put(channel.getHome() + "/" + AnsibleResources.COMMONS_CONFIG_FILE));
-
-            // @ToDo JK :: Support custom site file
-            // commonConfig.writeSiteFile(channel.put(channel.getHome() + "/" + AnsibleResources.SITE_CONFIG_FILE));
-            // Write slave instance specific configuration file
-            for (Instance slave : slaveInstances) {
-                String filename = channel.getHome() + "/" + AnsibleResources.CONFIG_ROOT_PATH + "/" + slave.getPrivateIp() + ".yml";
-                commonConfig.writeInstanceFile(slave, channel.put(filename));
-            }
-            uploadCompleted = true;
-        } catch (SftpException | IOException e) {
-            LOG.error("SFTP: {}", e);
-            uploadCompleted = false;
-        } finally {
-            channel.disconnect();
+        } catch (SftpException e) {
+            e.printStackTrace();
         }
-        return uploadCompleted;
     }
 
+    /**
+     * Creates folders for every directory, given a file structure.
+     *
+     * @param channel client side of sftp server channel
+     * @param resources ansible configuration
+     * @param files list of files in file structure
+     * @throws SftpException possible SFTP failure
+     */
     private void createSftpFolders(ChannelSftp channel, AnsibleResources resources, List<String> files) throws SftpException {
         for (String folderPath : resources.getDirectories(files)) {
             String fullPath = channel.getHome() + "/" + folderPath;
-            LOG.info(V, "SFTP: Create folder {}", fullPath);
             try {
                 channel.cd(fullPath);
             } catch (SftpException e) {
+                LOG.info(V, "SFTP: Create folder {}", fullPath);
                 channel.mkdir(fullPath);
             }
             channel.cd(channel.getHome());
         }
     }
 
-    private void uploadAnsibleRole(ChannelSftp channel, AnsibleResources resources, String rolePath, String roleName)
+    /**
+     * Uploads single ansible role (.tar.gz, .tgz) to remote instance.
+     *
+     * @param channel client side of sftp server channel
+     * @param resources ansible configuration
+     * @param roleFile path/to/role on local machine
+     * @throws SftpException possible SFTP failure
+     * @throws IOException possible File failure
+     */
+    private void uploadAnsibleRole(ChannelSftp channel, AnsibleResources resources, String roleFile)
             throws SftpException, IOException {
-        String basePath = AnsibleResources.ROLES_ROOT_PATH + "/" + roleName + "/";
-        Path rootRolePath = Paths.get(rolePath);
+        Path rootRolePath = Paths.get(roleFile);
+        // playbook/roles/ROLE_NAME
+        roleFile = this.getSingleFileName(roleFile);
+        String remotePath = AnsibleResources.ROLES_ROOT_PATH + roleFile;
+        // Walks through valid files in rootRolePath and collects Stream to path list
         List<Path> files = Files.walk(rootRolePath).filter(p -> p.toFile().isFile()).collect(Collectors.toList());
-        List<String> targetFiles = files.stream().map(p -> basePath + rootRolePath.relativize(p)).collect(Collectors.toList());
+        // create a list of files in role
+        List<String> targetFiles = files.stream().map(p -> remotePath + rootRolePath.relativize(p)).collect(Collectors.toList());
         createSftpFolders(channel, resources, targetFiles);
         for (int i = 0; i < files.size(); i++) {
             InputStream stream = new FileInputStream(files.get(i).toFile());
@@ -375,8 +496,28 @@ public abstract class CreateCluster extends Intent {
         }
     }
 
+    /**
+     * Turns path/to/file.* into file.*.
+     * @param roleFile path/to/file
+     * @return fileName
+     */
+    private String getSingleFileName(String roleFile) {
+        roleFile = roleFile.replace("\\", "/");
+        String[] pathway = roleFile.split("/");
+        return pathway[pathway.length - 1];
+    }
+
+    /**
+     * Installs and executes ansible roles on remote.
+     *
+     * @param sshSession transfer via ssh session
+     * @param prepare true, if still preparation necessary
+     * @return true, if configuration finished
+     * @throws JSchException ssh openChannel exception
+     * @throws IOException BufferedReader exceptions
+     */
     private boolean installAndExecuteAnsible(final Session sshSession, final boolean prepare)
-            throws JSchException, IOException {
+            throws IOException, JSchException {
         LOG.info("Configure and execute Ansible. This take a while. Please be patient.");
         boolean configured = false;
         String execCommand = ShellScriptCreator.getMasterAnsibleExecutionScript(prepare);